Move v2 annos to own namespace
