+++ src/Wellcome.Dds/IIIF/Presentation/V2/Annotation/Annotation.cs
@@ -1,7 +1,7 @@
 using IIIF.Presentation.V3.Annotation;
 using Newtonsoft.Json;
 
-namespace IIIF.Presentation.V2
+namespace IIIF.Presentation.V2.Annotation
 {
     public class Annotation : LegacyResourceBase, IAnnotation
     {
