+++ src/Wellcome.Dds/IIIF/Presentation/V2/Annotation/SearchResultAnnotation.cs
@@ -1,4 +1,4 @@
-namespace IIIF.Presentation.V2
+namespace IIIF.Presentation.V2.Annotation
 {
     public class SearchResultAnnotation : Annotation
     {
