+++ src/Wellcome.Dds/IIIF/Presentation/V2/Annotation/SearchResultAnnotationResource.cs
@@ -1,6 +1,6 @@
 using Newtonsoft.Json;
 
-namespace IIIF.Presentation.V2
+namespace IIIF.Presentation.V2.Annotation
 {
     public class SearchResultAnnotationResource : LegacyResourceBase
     {
