+++ src/Wellcome.Dds/Wellcome.Dds/IIIFBuilding/AltoAnnotationBuildResult.cs
@@ -1,4 +1,5 @@
 using IIIF.Presentation.V2;
+using IIIF.Presentation.V2.Annotation;
 using IIIF.Presentation.V3.Annotation;
 using Wellcome.Dds.AssetDomain.Mets;
